added additional abstraction layer for modulable storage multi
