+++ src/main/java/tkcy/simpleaddon/common/metatileentities/multiprimitive/MetaTileEntityMultiblockStorage.java
@@ -1,15 +1,12 @@
 package tkcy.simpleaddon.common.metatileentities.multiprimitive;
 
 import static gregtech.api.util.RelativeDirection.*;
-import static gregtech.api.util.RelativeDirection.UP;
 
 import java.util.List;
 import java.util.function.Function;
 
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.ResourceLocation;
@@ -19,30 +16,19 @@ import net.minecraft.util.text.Style;
 import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.util.text.event.HoverEvent;
-import net.minecraft.world.World;
 import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.fluids.FluidStack;
-import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.lwjgl.input.Keyboard;
 
-import gregtech.api.capability.impl.FilteredFluidHandler;
-import gregtech.api.capability.impl.FluidTankList;
-import gregtech.api.capability.impl.PropertyFluidFilter;
 import gregtech.api.metatileentity.MetaTileEntity;
-import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
 import gregtech.api.metatileentity.multiblock.IMultiblockPart;
 import gregtech.api.metatileentity.multiblock.MultiblockWithDisplayBase;
 import gregtech.api.pattern.BlockPattern;
 import gregtech.api.pattern.FactoryBlockPattern;
 import gregtech.api.pattern.PatternMatchContext;
 import gregtech.api.unification.material.Material;
-import gregtech.api.unification.material.properties.FluidPipeProperties;
-import gregtech.api.unification.material.properties.PropertyKey;
 import gregtech.client.renderer.ICubeRenderer;
 import gregtech.client.renderer.texture.Textures;
 
@@ -57,37 +43,33 @@ import tkcy.simpleaddon.api.metatileentities.MetaTileEntityStorageFormat;
 import tkcy.simpleaddon.api.metatileentities.RepetitiveSide;
 import tkcy.simpleaddon.api.predicates.TKCYSAPredicates;
 import tkcy.simpleaddon.api.render.TKCYSATextures;
-import tkcy.simpleaddon.api.utils.MaterialHelper;
 import tkcy.simpleaddon.api.utils.StorageUtils;
 import tkcy.simpleaddon.api.utils.units.CommonUnits;
 import tkcy.simpleaddon.api.utils.units.UnitsConversions;
 import tkcy.simpleaddon.common.block.TKCYSAMetaBlocks;
+import tkcy.simpleaddon.common.metatileentities.storage.MetaTileEntityModulableValve;
 import tkcy.simpleaddon.modules.storagemodule.StorageModule;
 
 @StorageModule.StorageModulable
-public class ModulableTank extends MultiblockWithDisplayBase
-                           implements RepetitiveSide, BlockMaterialMetaTileEntityPaint, MaterialMetaTileEntity,
-                           MetaTileEntityStorageFormat<FluidStack> {
+public abstract class MetaTileEntityMultiblockStorage<T, U> extends MultiblockWithDisplayBase
+                                                     implements RepetitiveSide, BlockMaterialMetaTileEntityPaint,
+                                                     MaterialMetaTileEntity,
+                                                     MetaTileEntityStorageFormat<U> {
 
     @Getter
     private final Material material;
     private int height;
-    private final boolean isLarge;
-    private final int layerCapacity;
-    private int totalCapacity;
-    private FluidPipeProperties fluidPipeProperties;
-    private FilteredFluidHandler tank;
+    protected final boolean isLarge;
+    protected int layerCapacity;
+    @Getter
+    protected int totalCapacity;
 
-    public ModulableTank(ResourceLocation metaTileEntityId, Material material, boolean isLarge) {
+    public MetaTileEntityMultiblockStorage(ResourceLocation metaTileEntityId, Material material, boolean isLarge) {
         super(metaTileEntityId);
         this.material = material;
         this.isLarge = isLarge;
-        this.layerCapacity = (int) Math.pow(10, 6) * (isLarge ? 21 : 1);
+        setLayerCapacity(isLarge);
         initializeInventory();
-    }
-
-    private void setFluidPipeProperties() {
-        this.fluidPipeProperties = MaterialHelper.getMaterialProperty(this.material, PropertyKey.FLUID_PIPE);
     }
 
     @Override
@@ -97,34 +79,9 @@ public class ModulableTank extends MultiblockWithDisplayBase
         this.totalCapacity = this.layerCapacity * this.height;
     }
 
-    @Override
-    protected void initializeInventory() {
-        if (this.material == null) return;
-        super.initializeInventory();
+    protected abstract void setLayerCapacity(boolean isLarge);
 
-        setFluidPipeProperties();
-
-        this.tank = new FilteredFluidHandler(this.totalCapacity);
-        this.tank.setFilter(new PropertyFluidFilter(
-                this.fluidPipeProperties.getMaxFluidTemperature(),
-                false,
-                this.fluidPipeProperties.isAcidProof(),
-                false,
-                false));
-
-        this.exportFluids = this.importFluids = new FluidTankList(true, this.tank);
-        this.fluidInventory = this.tank;
-    }
-
-    @Override
-    public MetaTileEntity createMetaTileEntity(IGregTechTileEntity tileEntity) {
-        return new ModulableTank(metaTileEntityId, material, isLarge);
-    }
-
-    @Override
-    protected void updateFormedValid() {
-        this.tank.setCapacity(this.totalCapacity);
-    }
+    protected abstract MetaTileEntityModulableValve<T> getValve(Material material);
 
     @NotNull
     @Override
@@ -136,7 +93,7 @@ public class ModulableTank extends MultiblockWithDisplayBase
                 .where('I', TKCYSAPredicates.isAir(RepetitiveSide.getHeightMarker()))
                 .where('X',
                         TKCYSAPredicates.iBlockStatePredicate(getSideBlockBlockState())
-                                .or(TKCYSAPredicates.metaTileEntityPredicate(StorageModule.getValve(this.material))
+                                .or(TKCYSAPredicates.metaTileEntityPredicate(getValve(this.material))
                                         .setMaxGlobalLimited(4)))
                 .build();
     }
@@ -207,11 +164,15 @@ public class ModulableTank extends MultiblockWithDisplayBase
         return Textures.MULTIBLOCK_TANK_OVERLAY;
     }
 
+    protected abstract Capability<T> getCapability();
+
+    protected abstract T getHandler();
+
     @Override
     public <T> T getCapability(Capability<T> capability, EnumFacing side) {
-        if (capability == CapabilityFluidHandler.FLUID_HANDLER_CAPABILITY) {
+        if (capability == getCapability()) {
             if (isStructureFormed()) {
-                return CapabilityFluidHandler.FLUID_HANDLER_CAPABILITY.cast(this.fluidInventory);
+                return getCapability().cast(getHandler());
             } else {
                 return null;
             }
@@ -259,26 +220,8 @@ public class ModulableTank extends MultiblockWithDisplayBase
     }
 
     @Override
-    public void addInformation(ItemStack stack, @Nullable World player, @NotNull List<String> tooltip,
-                               boolean advanced) {
-        tooltip.add(I18n.format("tkcysa.multiblock.modulable_tank.tooltip"));
-        tooltip.add(I18n.format(
-                "tkcysa.multiblock.modulable_storage.layer_infos", getCapacityPerLayerFormatted(), getMaxSideLength()));
-
-        if (Keyboard.isKeyDown(Keyboard.KEY_LSHIFT) || Keyboard.isKeyDown(Keyboard.KEY_RSHIFT)) {
-            tooltip.add(I18n.format("gregtech.fluid_pipe.max_temperature",
-                    this.fluidPipeProperties.getMaxFluidTemperature()));
-            if (this.fluidPipeProperties.isAcidProof()) {
-                tooltip.add(I18n.format("gregtech.fluid_pipe.acid_proof"));
-            }
-        } else {
-            tooltip.add(I18n.format("gregtech.tooltip.fluid_pipe_hold_shift"));
-        }
-    }
-
-    @Override
     public void displayInfos(List<ITextComponent> textList) {
-        StorageUtils<FluidStack> stackStorageUtil = getStorageUtil();
+        StorageUtils<U> stackStorageUtil = getStorageUtil();
 
         textList.add(stackStorageUtil.getCapacityTextTranslation());
         textList.add(stackStorageUtil.getContentTextTranslation());
@@ -288,16 +231,6 @@ public class ModulableTank extends MultiblockWithDisplayBase
     @Override
     public int getMaxCapacity() {
         return this.totalCapacity;
-    }
-
-    @Override
-    public Function<FluidStack, String> getContentLocalizedNameProvider() {
-        return FluidStack::getLocalizedName;
-    }
-
-    @Override
-    public Function<FluidStack, Integer> getContentAmountProvider() {
-        return fluidStack -> fluidStack.amount;
     }
 
     @Override
@@ -316,28 +249,7 @@ public class ModulableTank extends MultiblockWithDisplayBase
     }
 
     @Override
-    public StorageUtils<FluidStack> getStorageUtil() {
+    public StorageUtils<U> getStorageUtil() {
         return new StorageUtils<>(this);
-    }
-
-    @Override
-    @Nullable
-    public FluidStack getContent() {
-        return this.fluidInventory.drain(Integer.MAX_VALUE, false);
-    }
-
-    @Override
-    public String getPercentageTranslationKey() {
-        return "tkcysa.multiblock.modulable_storage.fill.percentage";
-    }
-
-    @Override
-    public String getCapacityTranslationKey() {
-        return "tkcysa.multiblock.modulable_storage.capacity";
-    }
-
-    @Override
-    public String getContentTextTranslationKey() {
-        return "tkcysa.multiblock.modulable_storage.content";
     }
 }
